requirements.txt
akismet==1.1
amqp==2.6.1
appdirs==1.4.4
asgiref==3.2.10
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
black==21.8b0
bleach==3.3.1
boto3==1.18.14
botocore==1.21.14
CacheControl==0.12.6
cachy==0.3.0
celery==4.3.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
cleo==0.8.1
click==8.0.1
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
clikit==0.6.2
coreapi==2.3.3
coreschema==0.0.4
coverage==5.3
coveralls==2.2.0
crashtest==0.3.1
cryptography==3.4.7
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.12
diff-match-patch==20200713
distlib==0.3.2
dj-database-url==0.5.0
Django==3.1
-e git+https://github.com/PetrKudy/django-admin-charts@e18c606d0d64822420e94b6db1bd3e12deec8c6f#egg=django_admin_charts
django-admin-large-initial-data==1.0.0
-e git+https://github.com/PetrKudy/django-admin-numeric-filter@57bfaba371f23fc65d62336d126e83c5ebd4a3e5#egg=django_admin_numeric_filter
django-admin-rangefilter==0.8.1
-e git+https://github.com/PetrDlouhy/django-admin-smoke-tests@dbf44b7e4a860490c6b48c1069a98a3b64bac7fb#egg=django_admin_smoke_tests_2
django-admin-tools==0.9.1
-e git+https://github.com/auto-mat/django-adminactions@0d744c91c2422a01ed561c2cc2efa331f3a4d827#egg=django_adminactions
django-adminfilters==1.0.0
-e git+https://github.com/tmszi/django-advanced-filters@1867d388a4d12ca5ef20657dc6a8d69a10dc8fd7#egg=django_advanced_filters
django-appconf==1.0.4
django-author==1.0.2
django-autoslug==1.9.8
-e git+https://github.com/jpic/django-betterforms@e392f496bbea8a3fd1b3f743376393ed7a8dd0b3#egg=django_betterforms
django-bootstrap-form==3.4
django-bootstrap4-form==4.0.1
django-bower==5.2.0
django-braces==1.14.0
django-bulk-update==2.2.0
django-celery-beat==2.0.0
django-celery-email==3.0.0
django-chart-tools==1.0
django-classy-tags==2.0.0
django-colorfield==0.3.2
django-computedfields==0.0.23
django-cors-headers==3.7.0
django-coverage-plugin==1.8.0
django-dbbackup==3.3.0
django-debug-toolbar==3.2.1
django-dynamic-fixture==3.1.1
django-extensions==3.1.3
django-extra-views==0.14.0
django-filter==2.4.0
-e git+https://github.com/auto-mat/django-flexible-filter-conditions@1d5ff9a587f2d63c0df36f0e5f92e22953a4f8b5#egg=django_flexible_filter_conditions
django-grapesjs==0.0.7
-e git+https://github.com/django-helpdesk/django-helpdesk@f53ee1366db1c98cc30140f1f3450751b3b9b4aa#egg=django_helpdesk
django-heroku==0.3.1
django-import-export==2.5.0
-e git+https://github.com/auto-mat/django-import-export-celery@d01672e4524c71e87b65c6603be196b2ffc18e07#egg=django_import_export_celery
django-inline-static==0.1.0
django-isnull-list-filter==0.2.0
django-js-asset==1.2.2
django-jsonfield==1.4.1
django-markdown-deux==1.0.5
django-mass-edit==3.2.0
django-model-utils==4.1.1
django-mptt==0.11.0
django-multiselectfield==0.1.12
django-nested-admin==3.3.3
django-notifications-hq==1.6.0
-e git+https://github.com/PetrKudy/django-nvd3@639300acdb014939c9f9562573f4b1425190724a#egg=django_nvd3
django-nyt==1.2
django-oauth-toolkit==1.5.0
django_polymorphic==3.1
-e git+https://github.com/Mogost/django-post_office@5297f00d886406de1233de6697262096e9ea81ca#egg=django_post_office
django-qsstats-magic==1.1.0
django-redis==5.0.0
django-related-admin==1.0.1
django-reversion==4.0.0
django-sekizai==2.0.0
django-select-multiple-field==0.4.2
django-ses==2.2.0
django-sesame==1.4
django-smmapdfs==1.2.0
django-spaghetti-and-meatballs==0.4.2
django-stdimage==5.3.0
django-stdnumfield==0.2.0
django-storages==1.11.1
django-timezone-field==4.2.1
django-tinymce==3.3.0
django-user-accounts==3.0.4
django-webtest==1.9.7
django-wysiwyg==0.8.0
djangorestframework==3.12.4
djcacheutils==3.0.0
docopt==0.6.2
docutils==0.17.1
drf-yasg==1.20.0
email-reply-parser==0.5.12
et-xmlfile==1.1.0
Faker==8.11.0
feedparser==6.0.8
filelock==3.0.12
flake8==3.8.4
flake8-blind-except==0.1.1
flake8-builtins==1.5.3
flake8-comprehensions==3.3.0
flake8-import-order==0.18.1
flake8-tidy-imports==4.1.0
flower==0.9.5
freezegun==1.1.0
future==0.18.2
gunicorn==20.1.0
hiredis==2.0.0
html2text==2020.1.16
html5lib==1.1
humanize==3.1.0
idna==3.2
importlib-metadata==4.6.3
importlib-resources==5.1.4
inflection==0.5.1
iniconfig==1.1.1
ipython==7.16.1
ipython-genutils==0.2.0
itypes==1.2.0
jdcal==1.4.1
jedi==0.17.2
jeepney==0.6.0
Jinja2==3.0.1
jmespath==0.10.0
jsonfield==3.1.0
jsonfield2==4.0.0
jwcrypto==1.0
keyring==21.8.0
kombu==4.6.11
lockfile==0.12.2
lxml==4.6.3
Markdown==3.2.2
markdown2==2.4.0
MarkupPy==1.14
MarkupSafe==2.0.1
mccabe==0.6.1
mock==4.0.2
model-mommy==2.0.0
msgpack==1.0.2
mypy-extensions==0.4.3
mysql==0.0.3
mysql-connector-python-rf==2.2.2
mysqlclient==2.0.3
newrelic==6.8.0.163
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.7
packaging==21.0
parso==0.7.1
pastel==0.2.1
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
pinax-invitations==7.0.0
pinax-teams==2.0.0
pkginfo==1.7.0
platformdirs==2.3.0
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==3.0.8
psycopg2==2.8.6
ptyprocess==0.6.0
pudb==2019.2
py==1.10.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.7.4
pylev==1.4.0
pyparsing==2.4.7
PyPDF2==1.26.0
pytest==6.2.4
pytest-cov==2.12.1
pytest-django==4.4.0
pytest-freezegun==0.4.2
python-crontab==2.5.1
python-dateutil==2.8.2
python-dotenv==0.19.0
python-memcached==1.59
python-monkey-business==1.0.0
python-nvd3==0.15.0
python-slugify==5.0.2
python-stdnum==1.11
python-utils==2.5.1
pytz==2021.1
PyYAML==5.4.1
raven==6.10.0
redis==3.5.3
regex==2021.8.28
reportlab==3.6.0
requests==2.26.0
requests-toolbelt==0.9.1
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
SecretStorage==3.3.1
selenium==3.141.0
setuptools-git==1.2
sgmllib3k==1.0.0
shellingham==1.4.0
simplejson==3.17.3
single-beat==0.4.2
six==1.16.0
sorl-thumbnail==12.7.0
soupsieve==2.2.1
sqlparse==0.4.2
swapper==1.1.2.post1
tablib==0.14.0
text-unidecode==1.3
toml==0.10.2
tomli==1.2.1
tomlkit==0.7.2
tornadis==0.8.1
tornado==5.1.1
traitlets==4.3.3
typed-ast==1.4.3
typing-extensions==3.10.0.0
unicodecsv==0.14.1
Unidecode==1.2.0
uritemplate==3.0.1
urllib3==1.26.6
urwid==2.1.2
vine==1.3.0
virtualenv==20.4.7
vokativ==1.2.1
waitress==2.0.0
wcwidth==0.2.5
webencodings==0.5.1
WebOb==1.8.7
WebTest==3.0.0
whitenoise==5.3.0
wiki==0.7.6
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.5.0
